// src/domain/interface/dish.interface.ts
export interface DishInterface {
  dish_id?: number | null;
  description?: string | null;
  availability?: string | null;
  name?: string | null;
  price?: number | null;
  category?: string | null;
  available?: boolean | null;
  image_url?: string | null;
  promotional_price?: number | null;
  restaurant_id?: number | null;
}